NrSpectrumPhy: Add sanity check to make sure that we have the SINR of the expected TB
  V2X example: Add NrV2xWestToEastHighway example
  V2xKpi: Silence UNCOND logging
  UeRlcRxOutputStats: Add class to write UE MAC RxRlcPduWithTxRnti to a DB
  V2xKpi: Add class to write NR V2X KPIs to a DB
  NrUeMac: Fire trace of RLC reception from MAC
  NrSpectrumPhy,NrUeMac: Use GetSeconds to convert time to ms
  NrSpectrumPhy: Convert the ABORT to continue statement
  NrSpectrumPhy: Update the logic to save error model output
  NrUeMac,NrUePhy,NrSlUePhySapUser: Add a separate container for RX dstL2Ids
  examples: Use V2V highway channel model in CttcNrV2xDemoSimple

+++ examples/nr-v2x-examples/ue-phy-pscch-rx-output-stats.cc
@@ -91,7 +91,7 @@ UePhyPscchRxOutputStats::WriteCache ()
     {
       sqlite3_stmt *stmt;
       m_db->SpinPrepare (&stmt, "INSERT INTO " + m_tableName + " VALUES (?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?);");
-      ret = m_db->Bind (stmt, 1, static_cast<double> (v.m_timeMs));
+      ret = m_db->Bind (stmt, 1, v.m_timeMs);
       NS_ABORT_UNLESS (ret);
       ret = m_db->Bind (stmt, 2, static_cast<uint32_t> (v.m_cellId));
       NS_ABORT_UNLESS (ret);
